Added forecastDeleteWebhookKey to forecastTimeEntriesDeleteWebhook endpoint and testForecastTimeEntriesDeleteWebhook test.

+++ src/main/kotlin/fi/metatavu/timebank/api/impl/TimeEntriesApi.kt
@@ -5,6 +5,7 @@ import fi.metatavu.timebank.api.impl.translate.TimeEntryTranslator
 import fi.metatavu.timebank.model.ForecastDeleteWebhookEvent
 import javax.enterprise.context.RequestScoped
 import fi.metatavu.timebank.spec.TimeEntriesApi
+import org.eclipse.microprofile.config.inject.ConfigProperty
 import java.time.LocalDate
 import java.util.*
 import javax.inject.Inject
@@ -15,6 +16,9 @@ import javax.ws.rs.core.Response
  */
 @RequestScoped
 class TimeEntriesApi: TimeEntriesApi, AbstractApi() {
+
+    @ConfigProperty(name = "webhook_key")
+    lateinit var forecastWebhookKey: String
 
     @Inject
     lateinit var timeEntryController: TimeEntryController
@@ -31,11 +35,13 @@ class TimeEntriesApi: TimeEntriesApi, AbstractApi() {
         return createNoContent()
     }
 
-    override suspend fun forecastTimeEntriesDeleteWebhook(forecastDeleteWebhookEvent: ForecastDeleteWebhookEvent): Response {
-        timeEntryController.deleteEntry(forecastId = forecastDeleteWebhookEvent.`object`!!.id)
+    override suspend fun forecastTimeEntriesDeleteWebhook(forecastDeleteWebhookEvent: ForecastDeleteWebhookEvent, forecastDeleteWebhookKey: String?): Response {
+            if (forecastDeleteWebhookKey != forecastWebhookKey) {return createUnauthorized(message = "Invalid key!")}
+                timeEntryController.deleteEntry(forecastId = forecastDeleteWebhookEvent.`object`!!.id)
+                return createNoContent()
 
-        return createNoContent()
     }
+
 
     override suspend fun listTimeEntries(personId: Int?, before: LocalDate?, after: LocalDate?, vacation: Boolean?): Response {
         loggedUserId ?: return createUnauthorized(message = "Invalid token!")
